Revert "Merge branch 'bmonik' of https://github.com/jokopurwanto/bca-mini-bank into bmonik"
This reverts commit fcda93c5508b42547b6dd42d1600ae7a40e8e626, reversing
changes made to bfe81932c6041a3bf82483918c7da2ac08d39c69.

+++ src/main/java/com/bca/minibank/mail/ContentEmailWestBankPKWT.java
@@ -92,7 +92,7 @@ public class ContentEmailWestBankPKWT {
 				"<p>   Telp : (021) 08123456789 </p>";
 	}
 	
-	public String formatRp(int nominal) {
+	public String formatRp(Double nominal) {
 		DecimalFormat kursIndonesia = (DecimalFormat) DecimalFormat.getCurrencyInstance();
 		DecimalFormatSymbols formatRp = new DecimalFormatSymbols();
 
@@ -104,7 +104,7 @@ public class ContentEmailWestBankPKWT {
 		return kursIndonesia.format(nominal);
 	}
 
-	public void getContentSuccessSetorTunai(int idTransaksi, Date tglPengajuan, String jnsTransaksi, String noRekAsal, String noRekTujuan,int nominal, String adminName) {
+	public void getContentSuccessSetorTunai(int idTransaksi, Date tglPengajuan, String jnsTransaksi, String noRekAsal, String noRekTujuan, Double nominal, String adminName) {
 		contentSubject = "Transaksi Setor Tunai Anda Berhasil";
 		contentType = "text/html";
 		contentHeader = "<h1>Website West Bank</h1>\r\n" + 
@@ -162,7 +162,7 @@ public class ContentEmailWestBankPKWT {
 				"<p>Jl. Slipi Jaya No. 1 Jakarta Barat</p>\r\n" + 
 				"<p>   Telp : (021) 08123456789 </p>";
 	}
-	public void getContentFailedSetorTunai(int idTransaksi, Date tglPengajuan, String jnsTransaksi, String noRekAsal, String noRekTujuan, int nominal, String adminName, String msg) {
+	public void getContentFailedSetorTunai(int idTransaksi, Date tglPengajuan, String jnsTransaksi, String noRekAsal, String noRekTujuan, Double nominal, String adminName, String msg) {
 		contentSubject = "Transaksi Setor Tunai Anda Gagal";
 		contentType = "text/html";
 		contentHeader = "<h1>Website West Bank</h1>\r\n" + 
@@ -220,5 +220,4 @@ public class ContentEmailWestBankPKWT {
 				"<p>Jl. Slipi Jaya No. 1 Jakarta Barat</p>\r\n" + 
 				"<p>   Telp : (021) 08123456789 </p>";
 	}
-
 }
